Implement #31: Drag and drop between boards
Cards and lanes can be dragged between board views.
(Note that dragging between boards can break links,
depending on your Obsidian link settings and whether
you have notes with the same name in different
locations.)

This is a significant re-organization and should
probably be tested more thoroughly, as I only use
a fraction of the plugin's features and might not
notice if I broke something.