Refresh react when workspace layout is reloaded
When the workspace is changed (or Obsidian starts), setViewData()
can be called before the view is mounted in the DOM, which prevents
the view's portal from rendering.  Once the workspace is fully
loaded, however, a change event is fired, allowing us to refresh
with the DOM actually in place.  (Other layout changes will be no-ops
since React is smart enough to know no props changed.)