Incremental parsing, reparse on settings change
This allows the board to be reparsed from on-disk changes
in 0.5-2ms per changed line.  Unfortunately, the rerendering
is still slow enough to make quick-preview impractical.  So
next up is checking on what else Item might be doing that forces
re-rendering, and how long Items take to render when they do.