Implement quick preview -- it's fast enough now
This will keep all open boards and editing views in sync.  Literally,
as in they will be synchronously updated.  This should make it
impossible to get the race condition invariant caused by dragging
one view of a file while the other view is trying to keep up.
(The key word there, of course, is *should*.)